clients: rename aiohttp_client_config to aiohttp_client
