+++ depobs/clients/npm_registry.py
@@ -5,7 +5,7 @@ import math
 from typing import Any, AsyncGenerator, Dict, Iterable, Optional, TypedDict
 import logging
 
-from depobs.clients.aiohttp_client_config import AIOHTTPClientConfig
+from depobs.clients.aiohttp_client import AIOHTTPClientConfig
 from depobs.util.type_util import Result
 from depobs.util.serialize_util import grouper
 from depobs.util.traceback_util import exc_to_str
